LayoutTests/accessibility/mac/focus-setting-selection-syncronizing-not-clearing-expected.txt


 1
 2
This tests that calling focus on a render object when it doesn't result in a selection change won't leave isSynchronizingSelection set to true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS axTextStateSyncOne is undefined
PASS accessibilityController.accessibleElementById("1").isFocusable is true
PASS accessibilityController.focusedElement.isEqual(accessibilityController.accessibleElementById("1")) is true
PASS axTextStateSyncTwo is undefined
PASS successfullyParsed is true

TEST COMPLETE
